+++ apps/frontend/components/DisasterTable/DetailedReportTable.tsx
@@ -1,14 +1,10 @@
 /* eslint-disable complexity */
-import {
-  GridColDef,
-  GridColumnGroupingModel,
-  GridSortCellParams,
-  gridStringOrNumberComparator,
-} from '@mui/x-data-grid';
+import { GridColDef, GridColumnGroupingModel } from '@mui/x-data-grid';
 import React from 'react';
 
 import { addDetailedReportLocationColumns } from 'utils/tableFormatting';
 
+import { createCustomComparator } from './CustomCommuneSort';
 import { DisasterTable } from './DisasterTable';
 
 interface IProps {
@@ -18,37 +14,6 @@ interface IProps {
   rotateHeader?: boolean;
   border?: boolean;
 }
-
-const getLocationIds = (id: string) => {
-  return {
-    provinceId: id.slice(0, 2),
-    districtId: id.slice(0, 4),
-  };
-};
-
-interface ExtendedGridSortCellParams extends GridSortCellParams {
-  id: string;
-  field: string;
-  value: string | number | undefined;
-}
-
-const findParentData = (
-  data: Record<string, string | number | undefined>[],
-  row: ExtendedGridSortCellParams,
-): {
-  provinceData: number | undefined;
-  districtData: number | undefined;
-} => {
-  const { provinceId, districtId } = getLocationIds(row.id);
-  const provinceData = data.find(
-    x => x.province === provinceId && x.district === undefined,
-  )?.[row.field] as number | undefined;
-  const districtData = data.find(
-    x => x.district === districtId && x.commune === undefined,
-  )?.[row.field] as number | undefined;
-
-  return { provinceData, districtData };
-};
 
 export const DetailedReportTable = ({
   columns,
@@ -120,62 +85,12 @@ export const DetailedReportTable = ({
     return provincesSummed.flat().flat();
   }, [data]);
 
-  // Custom comparator function
-
-  const isNumber = (
-    value: string | number | undefined,
-  ): value is number | undefined =>
-    typeof value === 'number' || value === undefined;
-
-  const customComparator = (
-    v1: number | undefined,
-    v2: number | undefined,
-    row1: ExtendedGridSortCellParams,
-    row2: ExtendedGridSortCellParams,
-  ) => {
-    if (!isNumber(row1.value) || !isNumber(row2.value)) {
-      return gridStringOrNumberComparator(v1, v2, row1, row2);
-    }
-
-    const parentData1 = findParentData(summedData, row1);
-    const parentData2 = findParentData(summedData, row2);
-
-    // Compare province data
-    if (parentData1.provinceData !== parentData2.provinceData) {
-      const result =
-        (parentData2.provinceData ?? 0) - (parentData1.provinceData ?? 0);
-
-      return result;
-    }
-
-    // Compare district data
-    if (parentData1.districtData !== parentData2.districtData) {
-      const result =
-        (parentData2.districtData ?? 0) - (parentData1.districtData ?? 0);
-
-      // Handle edge case for districtData
-      if (row1.id.length === 2 && !(row2.id.length === 2)) {
-        return -1; // Province rows should come before non-province rows
-      }
-      if (!(row1.id.length === 2) && row2.id.length === 2) {
-        return 1; // Non-province rows should come after province rows
-      }
-
-      return result;
-    }
-
-    // Compare row values
-    const result = (v2 ?? -Infinity) - (v1 ?? -Infinity);
-
-    return result;
-  };
-
   // Add custom comparator to relevant columns
   const columnsWithComparator = columns.map(col => {
     if (col.type === 'number') {
       return {
         ...col,
-        sortComparator: customComparator,
+        sortComparator: createCustomComparator(summedData),
       };
     }
